--- project/src/app/app-routing.module.ts
@@ -1,14 +1,0 @@
-import {NgModule} from "@angular/core";
-import {Routes, RouterModule} from "@angular/router";
-import {LoginComponent} from "./components/login/login.component";
-
-const routes: Routes = [
-  {path: '', pathMatch: 'full', component: LoginComponent},  
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule],
-  providers: []
-})
-export class AppRouterModule {}
