+++ project/src/app/app.component.html
@@ -1,5 +1,4 @@
 <navbar></navbar>
-<header></header>
 
 <!--<div class="container-fluid">
   <div class="row jumbotron">
