+++ project/src/app/app.component.ts
@@ -1,14 +1,14 @@
 import { Component, OnInit } from '@angular/core';
 import { FirebaseService } from './services/firebase.service';
 import { AngularFire, AuthMethods, AuthProviders } from 'angularfire2';
-import { AuthService } from './services/auth.service';
+import { Auth } from './services/auth.service';
 import 'rxjs/add/operator/map';
 
-// import { Course } from './models/Course';
-import { Category } from './models/Category';
-import { Status } from './models/Status';
-import { UserCourse } from './models/UserCourse';
-import { AllCourses } from "./models/AllCourses";
+// import { Course } from './firebase/Course';
+import { Category } from './firebase/Category';
+import { Status } from './firebase/Status';
+import { UserCourse } from './firebase/UserCourse';
+import { AllCourses } from "./firebase/AllCourses";
 
 @Component({
      selector: 'app-root',
@@ -18,7 +18,7 @@ import { AllCourses } from "./models/AllCourses";
 })
 
 export class AppComponent implements OnInit {
-    // isAuth = false;
+    isAuth = false;
     authColor = 'warn';
     user = {};
     allcourses: AllCourses[];
@@ -34,7 +34,7 @@ export class AppComponent implements OnInit {
     activeCategory: string;
     activeStatus: string;
     
-    constructor(private _firebaseService: FirebaseService, /* private auth: Auth, */ public af: AngularFire, private authService: AuthService
+    constructor(private _firebaseService: FirebaseService, private auth: Auth, public af: AngularFire
       ) {
         this.af.auth.subscribe(user => {
           if(user) {
@@ -107,12 +107,8 @@ export class AppComponent implements OnInit {
       
     login(from: string) {
       this.af.auth.login({
-        /*
         provider: AuthProviders.Google,
         method: AuthMethods.Popup
-        */
-        provider: AuthProviders.Password,
-        method: AuthMethods.Password
         // method: AuthMethods.Redirect
       });
       console.log("Logging In");
@@ -146,13 +142,5 @@ export class AppComponent implements OnInit {
     }
   }
 
-  isAuth() {
-    return this.authService.isAuthenticated();
-  }
-
-  onLogout() {
-    this.authService.logout();
-
-  }
 
 }
