+++ project/src/app/components/addusers/addusers.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from "@angular/core";
 import { FirebaseService } from "../../services/firebase.service";
-import { UserCourse } from '../../models/UserCourse';
-import { Category } from '../../models/Category';
-import { Status } from '../../models/Status';
+
+import { UserCourse } from '../../firebase/UserCourse';
+import { Category } from '../../firebase/Category';
+import { Status } from '../../firebase/Status';
 
 @Component({
