+++ project/src/app/components/header/header.component.html
@@ -1,40 +1,2 @@
 <h6>Header Goes Here</h6>
 <!-- <SlideShow></SlideShow> -->
-<!-- Facebook Code -->
-<script>
-!function(f,b,e,v,n,t,s){if(f.fbq)return;n=f.fbq=function(){n.callMethod?
-n.callMethod.apply(n,arguments):n.queue.push(arguments)};if(!f._fbq)f._fbq=n;
-n.push=n;n.loaded=!0;n.version='2.0';n.queue=[];t=b.createElement(e);t.async=!0;
-t.src=v;s=b.getElementsByTagName(e)[0];s.parentNode.insertBefore(t,s)}(window,
-document,'script','https://connect.facebook.net/en_US/fbevents.js');
-fbq('init', '295481444187735'); 
-fbq('track', 'PageView');
-</script>
-<noscript><img height="1" width="1" style="display:none"
-src="https://www.facebook.com/tr?id=295481444187735&ev=PageView&noscript=1"
-/></noscript>
-<!-- DO NOT MODIFY -->
-<!-- End Facebook Code -->
-
-<header>
-                  <nav class="navbar navbar-default">
-                                    <div class="container-fluid">
-                                                      <ul class="nav navbar-nav">
-                                                                        <li><a [routerLink]="['signup']">Sign Up</a></li>
-                                                                        <li><a [routerLink]="['signin']">Sign In</a></li>
-                                                                        <li><a [routerLink]="['showdata']">Show Data</a></li>
-                                                      </ul>
-                                                      <ul class="nav navbar-nav navbar-right" *ngIf="isAuth()">
-                                                                        <li><a (click)="onLogout()" style="cursor: pointer;">Logout</a></li>
-                                                      </ul>
-                                    </div>
-                  </nav>
-</header>
-
-<div class="container">
-  <div class="row">
-    <div class="col-md-10 col-md-offset-1">
-      <router-outlet></router-outlet>
-    </div>
-  </div>
-</div>
