+++ project/src/app/components/header/header.component.ts
@@ -6,7 +6,4 @@ import { Component } from '@angular/core';
   templateUrl: 'header.component.html'
 })
 
-export class HeaderComponent { 
-
-  
-}
+export class HeaderComponent { }
