+++ project/src/app/components/login/login.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { AuthService } from '../../services/auth.service';
+import { Auth } from '../../services/auth.service';
 import { Router } from '@angular/router';
 import { AngularFire, AuthProviders } from 'angularfire2';
