+++ project/src/app/components/testing/testing.component.ts
@@ -1,12 +1,12 @@
 import { Component } from '@angular/core';
-import { AuthService } from '../../services/auth.service';
+import { Auth } from '../../services/auth.service';
 import { FirebaseService } from '../../services/firebase.service';
 
 import { AddformComponent } from '../addform/addform.component';
-import { AllCourses } from '../../models/AllCourses';
-import { UserCourse } from '../../models/UserCourse';
-import { Category } from '../../models/Category';
-import { Status } from '../../models/Status';
+import { AllCourses } from '../../firebase/AllCourses';
+import { UserCourse } from '../../firebase/UserCourse';
+import { Category } from '../../firebase/Category';
+import { Status } from '../../firebase/Status';
 
 
 @Component({
@@ -16,6 +16,6 @@ import { Status } from '../../models/Status';
 })
 
 export class TestingComponent {
-  constructor(private auth: AuthService) {
+  constructor(private auth: Auth) {
   }
 }
