+++ project/src/app/components/welcome/welcome.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { AuthService } from '../../services/auth.service';
+import { Auth } from '../../services/auth.service';
 import { MdToolbarModule } from '@angular2-material/toolbar';
 import { MdButtonModule } from '@angular2-material/button';
 import { MdCardModule } from '@angular2-material/card';
@@ -11,7 +11,7 @@ import { MdCardModule } from '@angular2-material/card';
 })
 
 export class WelcomeComponent { 
-  constructor(private auth: AuthService) {
+  constructor(private auth: Auth) {
 
 
   }
