--- project/src/app/models/User.ts
@@ -1,4 +1,0 @@
-export interface User {
-    $key?: string;
-    name?: string;
-}
